Meaningfull message

+++ src/main/java/org/model/Address.java
@@ -9,7 +9,7 @@ public class Address extends Queryable{
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private long id;
 
     @Column(name = "country")
     private String country;
@@ -40,11 +40,11 @@ public class Address extends Queryable{
     }
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
